Rewrote the config code to use a crate instead
The configuration is now a lot clearer and is able to test-read a file. Just one
for now, I haven't extended it to multi-file support just yet -- but the path to
do so is quite clear. I should be able to do that in a pleasingly functional
manner, once I get to it.

This also removes a bunch of unnecessary types and otherwise neatens up the code.