Restored test files to TOML
Now the test files use TOML again, they're in their own directory, and multiple
files can be loaded at once and run. I'll add notes about what to expect or not
in terms of record names to the readme; one thing to note is that all record
names must be unique, and can't contain periods.