Adding detail to the README about configuration and plans
Giving something of a roadmap for future goals for this tool. I don't expect
this tool specifically to become a mechanism for, say, Let's Encrypt, but I
don't see why that should be outside of its purview, and if a record type can be
updated dynamically, I think this tool should be able to do it.