Added support for declaring a specific interface
Records now get updated over specific interfaces, and will only update a record
they're able to use an address to update. Note, this is potentially error-prone
if your network has link-local IPv6 but no route; I have not tried to insulate
this code to that extent. This code still expects some care on the part of the
user.